avoid remove&reinsert of call when splitting block for inlining
We used to split the inlined-into block at (= after) the call, and then
remove the call from the first block to insert it in the second.

The removal may cause unnecessary and unrecoverable resetting of debug
insns: we do not generate debug temps for calls.

Avoid the remove-and-reinsert dance by splitting the block before the
call.

for  gcc/ChangeLog

	* tree-inline.c (expand_call_inline): Split block at stmt
	before the call.

for  gcc/testsuite/ChangeLog

	* gcc.dg/guality/inline-params-2.c: New.

git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@247830 138bc75d-0d04-0410-961f-82ee72b054a4